fixed bug 2522 for learning modules
git-svn-id: http://svn.ilias.de/svn/ilias/trunk@13935 21b2c9ec-7c21-0410-8b45-9bfb8ed2bfc5

+++ Services/COPage/classes/class.ilPageObject.php
@@ -1262,8 +1262,11 @@ class ilPageObject
 		include_once("./Modules/File/classes/class.ilObjFile.php");
 		foreach ($files as $file_id)
 		{
-			$file_obj =& new ilObjFile($file_id, false);
-			$file_obj->delete();
+			if (ilObject::_exists($file_id))
+			{
+				$file_obj =& new ilObjFile($file_id, false);
+				$file_obj->delete();
+			}
 		}
 
 	}
